% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_homozygote_second_allele.R
\name{add_homozygote_second_allele}
\alias{add_homozygote_second_allele}
\title{Add second line to homozygote genotypes.}
\usage{
add_homozygote_second_allele(hap_raw)
}
\arguments{
\item{hap_raw}{is the raw (or lightly filtered) haplotypes you wish to add second
homozygote lines to}
}
\description{
Given a dataset (ideally the output from filter_raw_microhap_data), detect
individuals that have only one line at a locus and add a second line that is
a copy of the first with rank set to rank of original copy+1, so you have a
minimum of two lines (i.e. alleles) for each individual at each locus.
}
