% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_missing_data.R
\name{calculate_missing_data}
\alias{calculate_missing_data}
\title{Calculate number of loci with and without missing data.}
\usage{
calculate_missing_data(long_genos)
}
\arguments{
\item{long_genos}{genetic data in long format (can be filtered, unfiltered, and
with or without second homozygote alleles added, but MUST have the column names
from microhaplot).}
}
\description{
Calculate the number of loci total in the dataset (n_loci), the total number
of loci typed in each individual (n_typed), and the difference between the two
(n_miss).
}
