% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_bad_loci.R
\name{filter_bad_loci}
\alias{filter_bad_loci}
\title{Remove loci from data frame}
\usage{
filter_bad_loci(long_genos, bad_loci)
}
\arguments{
\item{long_genos}{genetic data}

\item{bad_loci}{character vector of locus names to remove}
}
\description{
Removes loci that the user determined were not useful, or
do not pass hardy-weinberg equilibrium, or you just don't
like them. Just be sure that manuscript reviewers will buy your
reason for removing these loci!
}
