% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_missing_data.R
\name{filter_missing_data}
\alias{filter_missing_data}
\title{Remove samples with too much missing data}
\usage{
filter_missing_data(long_genos, n_miss)
}
\arguments{
\item{long_genos}{A dataframe in long format (i.e. tidy format) that samples
are to be removed from. This could be the output dataframe from
'filter_raw_microhap_data'}

\item{n_miss}{The maximum number of loci with missing data for a sample
to be retained. Samples with more than 'n_miss' loci with missing
data are removed.}
}
\description{
This function removes samples from a data frame that
have too much missing data.
}
