% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_raw_microhap_data.R
\name{filter_raw_microhap_data}
\alias{filter_raw_microhap_data}
\title{Filter raw microhaplot output from read_unfiltered_observed().}
\usage{
filter_raw_microhap_data(hap_raw, haplotype_depth, total_depth, allele_balance)
}
\arguments{
\item{hap_raw}{dataframe of unfiltered haplotypes. This could be from a single run or a
large number of runs that have been joined together.}

\item{haplotype_depth}{remove haplotypes with less than depth specified here. This is for removing what's
considered potential genotyping errors or depths you don't feel confident about.}

\item{total_depth}{minimum number of reads required for the genotype to be retained.}

\item{allele_balance}{Ratio of the number of reads for a given microhaplotype divided
by the number of reads of the most common microhaplotype for that individual at that
locus.  For example, if you have a heterozygote with 60 reads for the haplotype AA and
100 reads for the haplotype TC, the allele balance for the haplotype AA for that individual
at that locus is 0.6 and is 1.0 for TC.}
}
\description{
Given a set of filtering criteria on individual haplotype depth, total read depth for
the genotype and allelle balance (the ratio of a particular microhaplotype read depth
to read depth of the most common microhaplotype for an individual at a locus) this
function will filter raw amplicon sequencing data. The input file here could be what
was loaded into R using the read_unfiltered_observed() function from this package.
}
