% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_missing_samples.R
\name{find_missing_samples}
\alias{find_missing_samples}
\title{Determine which samples did not pass filtering criteria}
\usage{
find_missing_samples(raw_data, filtered_data)
}
\arguments{
\item{raw_data}{The raw data from a single, or multiple sequencing runs. This could be the output from
read_unfiltered_observed()}

\item{filtered_data}{datafile of filtered haplotypes. This is the output from
the "filter_raw_microhap_data" function.}
}
\description{
Some samples may have failed completely and they will not be present in the
long_genos dataframe produced by filter_raw_microhap_data(). This function is
written to identify samples that were present in the raw data, but did not have
any loci that passed the filtering criteria. If a sample failed completely and did
not have any reads assigned to it, that sample will not be identified by this function.
}
