% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhap_transform.R
\name{mhap_transform}
\alias{mhap_transform}
\title{Transform long genos to different formats for other analysis programs}
\usage{
mhap_transform(long_genos, program)
}
\arguments{
\item{long_genos}{datafile of filtered haplotypes. This is the output from
the "filter_raw_microhap_data" function.}

\item{program}{The program you want to create an input file for. This could be CKMRsim, rubias,
franz, ADD OTHERS HERE.}
}
\description{
Takes a long genotype dataframe and transforms it to be the initial format
required for different genetic analysis programs. Current options for genetic
analysis programs include, CKMRsim, rubias, snppit(?), franz(?), adegenet,
structure (?), WHAT ELSE DO PEOPLE COMMONLY USE HERE....
}
