% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_run_statistics.R
\name{plot_run_statistics}
\alias{plot_run_statistics}
\title{Summarize data grouped by sequencing run.}
\usage{
plot_run_statistics(datafile, output_summary)
}
\arguments{
\item{datafile}{dataframe of microhaplotype data. This can be filtered or unfiltered data. For example,
unfiltered data such as hap_raw or filtered data such as long_genos can be used.}

\item{output_summary}{summary statistic to plot. Choose one of "total_reads_per_locus",
"total_reads_per_indiv" or "unique_haps_per_locus". Plotting all 3 summary statistics
in a single window is uninterpretable. So look at each summary statistic separately.}
}
\description{
Plot summaries of read depth per locus, reads per sample, and haplotypes per
locus by sequencing run. This function can be used for raw or filtered data, but
we think it would be most useful as a quick check of the filtered data before
proceeding to more advanced analyses or data transformations for different programs.
}
