% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_haplot_rds.R
\name{read_haplot_rds}
\alias{read_haplot_rds}
\title{Read raw microhaplot rds files from a directory or file.}
\usage{
read_haplot_rds(datapath)
}
\arguments{
\item{datapath}{path to microhaplot rds files. If this is a directory, then all the files within it
ending with \code{.rds} (but not '_posinfo.rds', or '_annotate.rds') will be processed. If \code{datapath}
is single file with an rds extension then only that file will be loaded. If it is  a vector of file paths,
at least one of which has a \code{.rds} extension, then the function will attempt to load all of the files.}
}
\description{
This function is designed to identify all raw rds data files within a folder and
join them together in R for further data processing. A single file or an entire
folder of data files can be read and joined with this function (note that this
will ignore the '_posinfo' and '_annotate' files generated by microhaplot be default).
If you only want a couple of runs read in rather than your whole microhaplot bin, or
just one run, you can feed this function a list of .rds files (just make sure they're
not '_posinfo' or '_annotate' files)
}
