% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_unfiltered_observed.R
\name{read_unfiltered_observed}
\alias{read_unfiltered_observed}
\title{Read unfiltered observed microhaplot output from a directory or file.}
\usage{
read_unfiltered_observed(datapath)
}
\arguments{
\item{datapath}{path to microhaplot output files. If a directory, then all csv files within it are
processed. If a single file, with a csv extension then only that file will be loaded.}
}
\description{
blah blah blah long descrip here
}
