% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_duplicate_samples.R
\name{resolve_duplicate_samples}
\alias{resolve_duplicate_samples}
\title{Resolves duplicate samples}
\usage{
resolve_duplicate_samples(long_genos)
}
\arguments{
\item{long_genos}{genetic data in long format where each sample
has 2 rows per locus. This dataframe could be the output from
'filter_raw_microhap_data'}
}
\description{
If a sample was sequenced on multiple sequencing runs (i.e.
it's present with multiple 'source' variables) this function
identifies the sequencing run with more missing data and removes
that sample-by-source combination. This function does not test
if the duplicates are genotypically identical.If you want to
confirm that they are truly duplicates with identical
genotype data go have a crack at 'CKMRsim::find_close_matching_genotypes'.
}
