% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_missing_data.R
\name{calculate_missing_data}
\alias{calculate_missing_data}
\title{Calculate number of loci with and without missing data.}
\usage{
calculate_missing_data(long_genos)
}
\arguments{
\item{long_genos}{genetic data in long format where each sample has 2 rows per locus.
the sample name must be in the 1st column of the data.}
}
\description{
Calculate the number of loci with missing data and the number
of loci with genotype data from a long format dataframe.
Both missing and non-missing loci are provided here to allow flexibility for the user.
}
