% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_raw_microhap_data.R
\name{filter_raw_microhap_data}
\alias{filter_raw_microhap_data}
\title{Filter raw microhaplot output from read_unfiltered_observed().}
\usage{
filter_raw_microhap_data(hap_raw, haplotype_depth, total_depth, allele_balance)
}
\arguments{
\item{hap_raw}{dataframe of unfiltered haplotypes. This could be from a single run or a
large number of runs that have been joined together.}

\item{haplotype_depth}{remove haplotypes with less than depth specified here. This is for removing what's
considered potential genotyping errors or depths you don't feel confident about.}

\item{total_depth}{minimum number of reads required for the genotype to be retained.}

\item{allele_balance}{For heterozygotes only. The ratio of the second haplotype read depth to the first
haplotype read depth. For example, genotype with read depth of 30 and 20, the allele balance is 0.66}
}
\description{
Given a set of filtering criteria on individual haplotype depth, total read depth for
the genotype and allelle balance (the ratio of read 2 depth to read 1 depth) this
function will filter raw amplicon sequencing data. The input file here could be
what was loaded into R using the read_unfiltered_observed() function from this package.
}
