% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_contaminated_samples.R
\name{find_contaminated_samples}
\alias{find_contaminated_samples}
\title{Find potentially contaminated samples}
\usage{
find_contaminated_samples(
  raw_data,
  haplotype_depth,
  total_depth,
  allele_balance
)
}
\arguments{
\item{raw_data}{The raw data from a single, or multiple sequencing runs. This could be the output from
read_unfiltered_observed()}
}
\description{
This function identifies samples that have more than 2 haplotypes per locus
that pass filtering on haplotype depth, total depth and allele balance.
}
