% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_run_statistics.R
\name{plot_run_statistics}
\alias{plot_run_statistics}
\title{Summarize data files grouped by group_id.}
\usage{
plot_run_statistics(datafile, num_runs)
}
\arguments{
\item{datafile}{dataframe of microhaplotype data. This can be filtered or unfiltered data. For example,
unfiltered data such as hap_raw or filtered data such as long_genos can be used.}

\item{num_runs}{The number of sequencing runs in the datafile.}
}
\description{
Plot summaries of read depth per locus, reads per sample, and haplotypes per
locus by sequencing run. This function can be used for raw or filtered data, but
we think it would be most useful as a quick check of the filtered data before
proceeding to more advanced analyses or data transformations for different programs.
}
