% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_unfiltered_observed.R
\name{read_unfiltered_observed}
\alias{read_unfiltered_observed}
\title{Read unfiltered observed microhaplot output from a directory or file.}
\usage{
read_unfiltered_observed(datapath, microhaplot_columns = TRUE)
}
\arguments{
\item{datapath}{path to microhaplot output files. If this is a directory, then all the files within it
ending with \code{.csv} or \code{.csv.gz} will be
processed. If \code{datapath} is single file with a csv extension then only that file will be loaded. If it is
a vector of file paths, at least one of which has a \code{.csv} or a \code{.csv.gz} extension, then
the function will attempt to load all of the files.}

\item{microhaplot_columns}{a logical.  Set to TRUE if you are reading output from microhaplot
and you can use FALSE if you are reading a file that you made yourself that might have
different column types than the microhaplot output.}
}
\description{
This function is designed to identify all raw data files within a folder and
join them together in R for further data processing. A single file or an entire
folder of data files can be read and joined with this function. If a subset of files
within a folder is desired, the user should supply the function with a list of file
paths.
}
