% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_unfiltered_observed.R
\name{read_unfiltered_observed}
\alias{read_unfiltered_observed}
\title{Read unfiltered observed microhaplot output from a directory or file.}
\usage{
read_unfiltered_observed(datapath)
}
\arguments{
\item{datapath}{path to microhaplot output files. If a directory, then all csv files within it are
processed. If a single file, with a csv extension then only that file will be loaded.}
}
\description{
This function is designed to identify all raw data files within a folder and
join them together in R for further data processing. A single file or an entire
folder of data files can be read and joined with this function. If a subset of files
within a folder is desired, the user should supply the function with a list of file
paths.
}
