% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_duplicate_samples.R
\name{resolve_duplicate_samples}
\alias{resolve_duplicate_samples}
\title{Resolves duplicate samples}
\usage{
resolve_duplicate_samples(haplo_data, resolve)
}
\arguments{
\item{haplo_data}{genetic data output from 'filter_raw_microhap_data'}

\item{resolve}{Enter "rename" if you wish to keep both runs of this sample and
"drop" if you want to drop the run with fewer loci genotyped.  The "rename"
option will sort your reruns by their "group" column in your datset and append
an "_1" to the first run's name, "_2" to the next, and so on.  The "drop" option
will not rename your samples, simply drop the run with fewer loci genotyped--if
they are equal, it will keep whichever one appears first in your dataset.}
}
\description{
If a sample was sequenced on multiple sequencing runs (i.e.
it's present with multiple 'source' variables) this function
identifies the sequencing run with more missing data and removes
that sample-by-source combination. This function does not test
if the duplicates are genotypically identical.If you want to
confirm that they are truly duplicates with identical
genotype data go have a crack at 'CKMRsim::find_close_matching_genotypes'.
}
